Drop autotools support

+++ PKGBUILD
@@ -17,27 +17,27 @@ optdepends=('socat: xdg-open client script support')
   sha256sums=('SKIP')
 }
 
+_make_flags=(PREFIX=/usr SYSCONFDIR=/etc)
+
 prepare() {
-  [ -n "${XOS_LOCAL}" ] && {
+  if [ -n "${XOS_LOCAL}" ]; then
     rm -rf "$srcdir/$pkgname"
     mkdir "$srcdir/$pkgname"
 
     (cd "$startdir" && find . -mindepth 1 -maxdepth 1 \
-    -not -name '.git' -and -not -name 'src' -print0) |
+    -not -name '.git' -and -not -name 'src' -and -not -name 'pkg' -print0) |
     xargs -0 -n 1 -I {} cp -rpv "$startdir/{}" "$srcdir/$pkgname/{}"
-  }
-
-  cd "$srcdir/$pkgname"
-  ./autogen.sh
+  else
+    true
+  fi
 }
 
 build() {
   cd "$srcdir/$pkgname"
-  ./configure --prefix=/usr --sysconfdir=/etc
-  make
+  make "${_make_flags[@]}"
 }
 
 package() {
   cd "$srcdir/$pkgname"
-  make DESTDIR="$pkgdir" install
+  make "${_make_flags[@]}" DESTDIR="$pkgdir" install
 }
